+++ inst/doc/design.R
@@ -6,9 +6,9 @@ knitr::opts_chunk$set(
 
 ## ------------------------------------------------------------------------
 (controlTable <- wrapr::qchar_frame(
-  flower_part, Length      , Width       |
-  Petal      , Petal.Length, Petal.Width |
-  Sepal      , Sepal.Length, Sepal.Width ))
+  "flower_part", "Length"    , "Width"     |
+  "Petal"      , Petal.Length, Petal.Width |
+  "Sepal"      , Sepal.Length, Sepal.Width ))
 
 (columnsToCopy <- "Species")