+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# cdata 1.0.8 2019/03/30
+
+ * More column collision checks.
+ * Operator notation.
+
 # cdata 1.0.7 2019/03/23
 
  * Move to wrapr tests.